+++ docs/html/search/all_5.js
@@ -1,19 +1,4 @@
 var searchData=
 [
-  ['getbuttonconfig',['getButtonConfig',['../classace__button_1_1AceButton.html#af5e39bdfbc0e6bbd9c9191029d29685e',1,'ace_button::AceButton']]],
-  ['getclickdelay',['getClickDelay',['../classace__button_1_1ButtonConfig.html#a19db203ce87dc2cafb2cab6ba124387a',1,'ace_button::ButtonConfig']]],
-  ['getclock',['getClock',['../classace__button_1_1ButtonConfig.html#a440271e22d5e7e161cdcb3b73f092111',1,'ace_button::ButtonConfig']]],
-  ['getdebouncedelay',['getDebounceDelay',['../classace__button_1_1ButtonConfig.html#ad7be58d0fe0b242f78f15319d201a565',1,'ace_button::ButtonConfig']]],
-  ['getdefaultreleasedstate',['getDefaultReleasedState',['../classace__button_1_1AceButton.html#afa8013bb98f5fbd79654aebca613ba71',1,'ace_button::AceButton']]],
-  ['getdoubleclickdelay',['getDoubleClickDelay',['../classace__button_1_1ButtonConfig.html#a5ba500367e03c36f39225a5858e110e6',1,'ace_button::ButtonConfig']]],
-  ['geteventhandler',['getEventHandler',['../classace__button_1_1ButtonConfig.html#af54641e72e5624be0eb57a012c5372bd',1,'ace_button::ButtonConfig']]],
-  ['getid',['getId',['../classace__button_1_1AceButton.html#ae81fb47e01b375ce1d79de3c8496ff21',1,'ace_button::AceButton']]],
-  ['getlastbuttonstate',['getLastButtonState',['../classace__button_1_1AceButton.html#acbb18322e7560cd526061918c50d462f',1,'ace_button::AceButton']]],
-  ['getlongpressdelay',['getLongPressDelay',['../classace__button_1_1ButtonConfig.html#a46d4213a827a94bf82ba99a9c2e59531',1,'ace_button::ButtonConfig']]],
-  ['getnobuttonpin',['getNoButtonPin',['../classace__button_1_1EncodedButtonConfig.html#aba96c8f8d55666fb413195407ec5770a',1,'ace_button::EncodedButtonConfig::getNoButtonPin()'],['../classace__button_1_1LadderButtonConfig.html#a1840f50aa3c070c5f0dc7957884c0643',1,'ace_button::LadderButtonConfig::getNoButtonPin()']]],
-  ['getpin',['getPin',['../classace__button_1_1AceButton.html#a64376f91e9053abb2a2040dd3c41daeb',1,'ace_button::AceButton']]],
-  ['getrepeatpressdelay',['getRepeatPressDelay',['../classace__button_1_1ButtonConfig.html#ad85c75d2a2dd200552a181fb154d5fd2',1,'ace_button::ButtonConfig']]],
-  ['getrepeatpressinterval',['getRepeatPressInterval',['../classace__button_1_1ButtonConfig.html#aec4a97a1a1306b9ff12f1cdf44515c43',1,'ace_button::ButtonConfig']]],
-  ['getsystembuttonconfig',['getSystemButtonConfig',['../classace__button_1_1ButtonConfig.html#a585353712d2c76d91c10e67102cb614d',1,'ace_button::ButtonConfig']]],
-  ['getvirtualpin',['getVirtualPin',['../classace__button_1_1EncodedButtonConfig.html#ae92e77f9b6e8f5a18cc4f0cf17406724',1,'ace_button::EncodedButtonConfig::getVirtualPin()'],['../classace__button_1_1LadderButtonConfig.html#aac2d59d472aadad37efa9a9c5509dc71',1,'ace_button::LadderButtonConfig::getVirtualPin()']]]
+  ['featureflagtype_13',['FeatureFlagType',['../classace__button_1_1ButtonConfig.html#a1c7c7bfc42738278330243dfc43f23eb',1,'ace_button::ButtonConfig']]]
 ];