+++ docs/html/search/functions_5.js
@@ -1,7 +1,19 @@
 var searchData=
 [
-  ['init',['init',['../classace__button_1_1AceButton.html#a18c47304c694c6f084a343a7c83bef34',1,'ace_button::AceButton::init(uint8_t pin=0, uint8_t defaultReleasedState=HIGH, uint8_t id=0)'],['../classace__button_1_1AceButton.html#a6433836fba70c590917284ff43e4afb4',1,'ace_button::AceButton::init(ButtonConfig *buttonConfig, uint8_t pin=0, uint8_t defaultReleasedState=HIGH, uint8_t id=0)']]],
-  ['isfeature',['isFeature',['../classace__button_1_1ButtonConfig.html#a23ca62772029a755b083ff4c34a27958',1,'ace_button::ButtonConfig']]],
-  ['ispressedraw',['isPressedRaw',['../classace__button_1_1AceButton.html#ae2409ef5624cd06b6465372562e2567e',1,'ace_button::AceButton']]],
-  ['isreleased',['isReleased',['../classace__button_1_1AceButton.html#aff39f5dff72b48718aed7052ab8e04af',1,'ace_button::AceButton']]]
+  ['getbuttonconfig_91',['getButtonConfig',['../classace__button_1_1AceButton.html#a7d309d463b50a59a202067a7228149f0',1,'ace_button::AceButton']]],
+  ['getclickdelay_92',['getClickDelay',['../classace__button_1_1ButtonConfig.html#a2a39fcc939e2dce12f2ddbf803747a6e',1,'ace_button::ButtonConfig']]],
+  ['getclock_93',['getClock',['../classace__button_1_1ButtonConfig.html#a440271e22d5e7e161cdcb3b73f092111',1,'ace_button::ButtonConfig']]],
+  ['getdebouncedelay_94',['getDebounceDelay',['../classace__button_1_1ButtonConfig.html#a2866cf111973f8b1bf4ae0e7df17b6dd',1,'ace_button::ButtonConfig']]],
+  ['getdefaultreleasedstate_95',['getDefaultReleasedState',['../classace__button_1_1AceButton.html#a5c253900d6d6c0035095a322ee7f9aa1',1,'ace_button::AceButton']]],
+  ['getdoubleclickdelay_96',['getDoubleClickDelay',['../classace__button_1_1ButtonConfig.html#a0098549db5df3c730f64f4235845a5a5',1,'ace_button::ButtonConfig']]],
+  ['geteventhandler_97',['getEventHandler',['../classace__button_1_1ButtonConfig.html#af0e25e21ee86af10e00e644c9be340a7',1,'ace_button::ButtonConfig']]],
+  ['getid_98',['getId',['../classace__button_1_1AceButton.html#a244a2ea3166f2d886ccd94a44286b835',1,'ace_button::AceButton']]],
+  ['getlastbuttonstate_99',['getLastButtonState',['../classace__button_1_1AceButton.html#abb9c10c1fc6df01fae071ccfcb906935',1,'ace_button::AceButton']]],
+  ['getlongpressdelay_100',['getLongPressDelay',['../classace__button_1_1ButtonConfig.html#a3ded707bd12aaf7bbcb14e431ec77e15',1,'ace_button::ButtonConfig']]],
+  ['getnobuttonpin_101',['getNoButtonPin',['../classace__button_1_1EncodedButtonConfig.html#aba96c8f8d55666fb413195407ec5770a',1,'ace_button::EncodedButtonConfig::getNoButtonPin()'],['../classace__button_1_1LadderButtonConfig.html#a1840f50aa3c070c5f0dc7957884c0643',1,'ace_button::LadderButtonConfig::getNoButtonPin()']]],
+  ['getpin_102',['getPin',['../classace__button_1_1AceButton.html#ab33671acf4de4912b1f00fad28246ae0',1,'ace_button::AceButton']]],
+  ['getrepeatpressdelay_103',['getRepeatPressDelay',['../classace__button_1_1ButtonConfig.html#aeca9329775d17a06433c9588fca28695',1,'ace_button::ButtonConfig']]],
+  ['getrepeatpressinterval_104',['getRepeatPressInterval',['../classace__button_1_1ButtonConfig.html#a5c5d86932a98f71030422ef367b57a25',1,'ace_button::ButtonConfig']]],
+  ['getsystembuttonconfig_105',['getSystemButtonConfig',['../classace__button_1_1ButtonConfig.html#a585353712d2c76d91c10e67102cb614d',1,'ace_button::ButtonConfig']]],
+  ['getvirtualpin_106',['getVirtualPin',['../classace__button_1_1EncodedButtonConfig.html#ae92e77f9b6e8f5a18cc4f0cf17406724',1,'ace_button::EncodedButtonConfig::getVirtualPin()'],['../classace__button_1_1LadderButtonConfig.html#aac2d59d472aadad37efa9a9c5509dc71',1,'ace_button::LadderButtonConfig::getVirtualPin()']]]
 ];